requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
camel-converter==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
instaloader==4.14.1
MarkupSafe==3.0.2
meilisearch==0.34.1
meilisearch-python-async==1.8.1
packaging==25.0
pillow==11.2.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
tldextract==5.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
yt-dlp==2025.3.31
